Fixed bogus error in clean repositories and removed hard reset instruction.
It seems like git should be sufficiently smart to understand which commits are already merged into the repository and thus we don't need to reset the repository back to its original state (this will also ensure that no commits are lost as a result of invoking a build without previously generating its patches). This change obviously also gets rid of the left over TODO comment within the safeguard mojo.